refactor: extract some PromQL Engine tests which use unexported structs
