+++ promql/engine_test.go
@@ -23,8 +23,6 @@ import (
 	"testing"
 	"time"
 
-	"github.com/go-kit/log"
-
 	"github.com/stretchr/testify/require"
 	"go.uber.org/goleak"
 
@@ -1712,63 +1710,6 @@ load 1ms
 			testutil.RequireEqual(t, c.result, res.Value, "query %q failed", c.query)
 		})
 	}
-}
-
-func TestRecoverEvaluatorRuntime(t *testing.T) {
-	var output []interface{}
-	logger := log.Logger(log.LoggerFunc(func(keyvals ...interface{}) error {
-		output = append(output, keyvals...)
-		return nil
-	}))
-	ev := &evaluator{logger: logger}
-
-	expr, _ := parser.ParseExpr("sum(up)")
-
-	var err error
-
-	defer func() {
-		require.EqualError(t, err, "unexpected error: runtime error: index out of range [123] with length 0")
-		require.Contains(t, output, "sum(up)")
-	}()
-	defer ev.recover(expr, nil, &err)
-
-	// Cause a runtime panic.
-	var a []int
-	a[123] = 1
-}
-
-func TestRecoverEvaluatorError(t *testing.T) {
-	ev := &evaluator{logger: log.NewNopLogger()}
-	var err error
-
-	e := errors.New("custom error")
-
-	defer func() {
-		require.EqualError(t, err, e.Error())
-	}()
-	defer ev.recover(nil, nil, &err)
-
-	panic(e)
-}
-
-func TestRecoverEvaluatorErrorWithWarnings(t *testing.T) {
-	ev := &evaluator{logger: log.NewNopLogger()}
-	var err error
-	var ws annotations.Annotations
-
-	warnings := annotations.New().Add(errors.New("custom warning"))
-	e := errWithWarnings{
-		err:      errors.New("custom error"),
-		warnings: warnings,
-	}
-
-	defer func() {
-		require.EqualError(t, err, e.Error())
-		require.Equal(t, warnings, ws, "wrong warning message")
-	}()
-	defer ev.recover(nil, &ws, &err)
-
-	panic(e)
 }
 
 func TestSubquerySelector(t *testing.T) {
